Refactor request routing
