+++ site/src/lib.rs
@@ -18,5 +18,6 @@ pub mod server;
 mod average;
 mod comparison;
 mod interpolate;
+mod request_handlers;
 mod selector;
 mod self_profile;
